change: change socket functions
