+++ src/components/profile/profile.js
@@ -1,5 +1,4 @@
 import { globalState, userState } from '../../../lib/state/state.js';
-import { listenFriendLogin } from './loginStatus.js';
 import { changeDateFormat } from '../../utils/date.js';
 import { escapeHtml } from '../../utils/validateInput.js';
 import { changeUserImageModal } from './modal/changeUserImage.js';
@@ -212,7 +211,6 @@ export async function setFriendList() {
   }
 
   friendList.innerHTML = '';
-  let friendIdArray = [];
 
   // listenFriendLogin();
   if (!friendData.friends.length) {
@@ -225,6 +223,7 @@ export async function setFriendList() {
     friendList.appendChild(friendItem);
   } else {
     const firstTwoResults = friendData.friends.slice(0, 2);
+    let friendIdArray = [];
 
     firstTwoResults.forEach(async (result) => {
       const friendItem = document.createElement('li');
@@ -285,8 +284,46 @@ export async function setFriendList() {
       friendItem.appendChild(friendListItemDiv);
       friendList.appendChild(friendItem);
     });
+    listenFriendLogin(friendIdArray);
   }
-  listenFriendLogin(friendIdArray);
+}
+
+function listenFriendLogin(array) {
+  const userData = userState.getState();
+  const userSocket = userData.userSocket;
+
+  const waitForSocketOpen = new Promise((resolve, reject) => {
+    const checkInterval = setInterval(() => {
+      if (userSocket && userSocket.readyState === WebSocket.OPEN) {
+        clearInterval(checkInterval);
+        resolve(userSocket);
+      } else if (userState.getState().socketStatus === 'offline') {
+        clearInterval(checkInterval);
+        reject(new Error('Socket is offline'));
+      }
+    }, 1000);
+  });
+
+  waitForSocketOpen
+    .then(() => {
+      const checkLoginInterval = setInterval(() => {
+        userSocket.send(JSON.stringify({ userid: array }));
+      }, 1000);
+
+      if (userState.getState().socketFunction) {
+        const previousInterval = userState.getState().socketFunction;
+        clearInterval(previousInterval);
+      } else {
+        userState.setState({ socketFunction: checkLoginInterval }, false);
+      }
+    })
+    .catch(() => {
+      const allLoginStatus = document.querySelectorAll('.login-status');
+      allLoginStatus.forEach((loginStatus) => {
+        loginStatus.classList.remove('logout');
+        loginStatus.classList.add('offline');
+      });
+    });
 }
 
 export async function setRequestList() {
@@ -467,11 +504,102 @@ function set2fa() {
   });
 }
 
+function listenSocketMessage() {
+  const userData = userState.getState();
+  const userSocket = userData.userSocket;
+
+  const waitForSocketOpen = new Promise((resolve, reject) => {
+    const checkInterval = setInterval(() => {
+      if (userSocket && userSocket.readyState === WebSocket.OPEN) {
+        console.log('socket open');
+
+        clearInterval(checkInterval);
+        resolve(userSocket);
+      } else if (userState.getState().socketStatus === 'offline') {
+        clearInterval(checkInterval);
+        reject(new Error('Socket is offline'));
+      }
+    }, 1000);
+  });
+
+  waitForSocketOpen
+    .then(() => {
+      userSocket.onmessage = (event) => {
+        const loginStatusList = JSON.parse(event.data);
+        // console.log('list: ', loginStatusList);
+
+        Object.entries(loginStatusList).forEach(([userId, isLoggedIn]) => {
+          const profileModal = globalState.getState().profileModal;
+          const viewAllModal = globalState.getState().viewAllModal;
+
+          if (!viewAllModal && !profileModal) {
+            // friends
+            const friendItem = document.getElementById(userId);
+            if (!friendItem) return;
+
+            const loginStatusDiv = friendItem.querySelector('.login-status');
+            if (isLoggedIn) {
+              loginStatusDiv.classList.remove('logout');
+              loginStatusDiv.classList.add('login');
+            } else {
+              loginStatusDiv.classList.remove('login');
+              loginStatusDiv.classList.add('logout');
+            }
+          }
+
+          if (profileModal) {
+            // profile modal
+            const modalElement = profileModal.modalInstance._element;
+            const modalUserId = profileModal.userId;
+
+            if (modalElement && modalUserId === +userId) {
+              const modalLoginStatusDiv =
+                modalElement.querySelector('.login-status');
+              if (isLoggedIn) {
+                modalLoginStatusDiv.classList.remove('logout');
+                modalLoginStatusDiv.classList.add('login');
+              } else {
+                modalLoginStatusDiv.classList.remove('login');
+                modalLoginStatusDiv.classList.add('logout');
+              }
+            }
+          }
+
+          if (viewAllModal) {
+            // view all modal
+            const modalElement = viewAllModal.modalInstance._element;
+            const modalItem = modalElement.querySelector(`[id="${userId}"]`);
+            if (!modalItem) return;
+
+            const modalLoginStatusDiv = modalItem.querySelector(
+              '.modal-login-status'
+            );
+            if (isLoggedIn) {
+              modalLoginStatusDiv.classList.remove('modal-logout');
+              modalLoginStatusDiv.classList.add('modal-login');
+            } else {
+              modalLoginStatusDiv.classList.remove('modal-login');
+              modalLoginStatusDiv.classList.add('modal-logout');
+            }
+          }
+        });
+      };
+    })
+    .catch(() => {
+      const allLoginStatus = document.querySelectorAll('.login-status');
+      allLoginStatus.forEach((loginStatus) => {
+        loginStatus.classList.remove('logout');
+        loginStatus.classList.add('offline');
+      });
+    });
+}
+
 export function profile() {
   setProfile();
   setHistoryList();
   setFriendList();
   setRequestList();
+  listenSocketMessage();
   setLanguage();
   set2fa();
   setModal();
